requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
bcrypt==5.0.0
blinker==1.9.0
boto3==1.42.30
botocore==1.42.30
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
google-genai==1.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
minio==7.2.20
narwhals==2.14.0
numpy==2.0.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==11.3.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.45
starlette==0.49.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
uvicorn==0.39.0
websockets==15.0.1
xlsxwriter==3.2.9
